functions/average/requirements.txt
google-cloud-bigquery==3.5.0
pydantic==1.10.5